Add response function viewer - intended to be data editor
The next step is to add the ability to select the frequencies within
the region as toggled masked or not with x for xy mode and y for yx
mode.

This should affect only the "maskeddata" dictionary, and the data
in there should be plotted as lines. The symbols should stay for all
data.